+++ src/commands/mover.js
@@ -1,13 +1,22 @@
 #!/usr/bin/env node
 const fs = require('fs')
-const { exec, execSync } = require('child_process')
-
-const { Params, PackageList, isWindows, redText, yellowText, greenText } = require('../utils')
+const path = require('path');
+const {
+    Params,
+    PackageList,
+    isWindows,
+    redText,
+    getDependencies,
+    resolveSubDependencies,
+    getPackageLock,
+    updatePackageJson,
+    execProm
+} = require('../utils');
 
 let params;
 
 function parseParams(args) {
-    if(args.alreadyParsed) return new Params(args.packages, args.source, args.destination, args.logErrors)
+    if (args.alreadyParsed) return new Params(args.packages, args.source, args.destination, args.logErrors)
     const p = new Params();
     for (let arg of args) {
         const [key, value] = arg.split("=");
@@ -23,30 +32,6 @@ function parseParams(args) {
     return p;
 }
 
-async function getDependencies(package) {
-    let packages = new PackageList();
-    try {
-        let {
-            dependencies,
-            devDependencies,
-            peerDependencies,
-            optionalDependencies
-        } = JSON.parse(fs.readFileSync(`${params.source}/node_modules/${package}/package.json`));
-
-        packages.add(
-            [
-                ...((dependencies && Object.keys(dependencies).map(el => ({ name: el, category: "dependencies" }))) ?? []),
-                ...((devDependencies && Object.keys(devDependencies).map(el => ({ name: el, category: "devDependencies" }))) ?? []),
-                ...((peerDependencies && Object.keys(peerDependencies).map(el => ({ name: el, category: "peerDependencies" }))) ?? []),
-                ...((optionalDependencies && Object.keys(optionalDependencies).map(el => ({ name: el, category: "optionalDependencies" }))) ?? [])
-            ]
-        )
-
-    } catch (e) {
-        if (params.logErrors) redText("[Error] Failed to resolve :", package);
-    }
-    return packages;
-}
 
 /**
  * @typedef {Object} StartParams
@@ -61,105 +46,105 @@ async function getDependencies(package) {
  * @param {StartParams} options 
  */
 async function start(options) {
-    
+    console.time('duration:')
     params = parseParams(options ?? process.argv.slice(2));
+    const packageLock = await getPackageLock(params.source, params)
     const packages = new PackageList();
+
     for (const pkg of params.packages) {
-        packages.add((await getDependencies(pkg)).values());
+        packages.add((await getDependencies(pkg, params, packageLock, true)).values());
     }
+
     for (const package_name of packages.names()) {
-        await resolveSubDependencies(package_name, packages)
+        await resolveSubDependencies(package_name, packages, params, packageLock)
     }
 
     console.log("Copying", packages.size(), "packages...")
 
-    copyFolders(packages.names())
-    updatePackageJson()
+    await copyFolders(packages.names())
+    await updatePackageJson(params, packages);
+    console.timeEnd('duration:')
 }
 
-async function resolveSubDependencies(package, mainPackages) {
-    const subpackages = await getDependencies(package);
-    for (const subpackage of subpackages.values()) {
-        if (!mainPackages.contains(subpackage.name)) {
-            mainPackages.add(subpackage)
-            await resolveSubDependencies(subpackage.name, mainPackages);
-        }
-        mainPackages.add(subpackage)
-    }
-}
+// async function updatePackageJson() {
+//     const source = JSON.parse(fs.readFileSync(`${params.source}/package.json`));
+//     const destination = JSON.parse(fs.readFileSync(`${params.destination}/package.json`));
 
-async function updatePackageJson() {
-    const source = JSON.parse(fs.readFileSync(`${params.source}/package.json`));
-    const destination = JSON.parse(fs.readFileSync(`${params.destination}/package.json`));
+//     let updatedDeps = {};
 
-    let updatedDeps = {};
+//     for (const package of params.packages) {
+//         const dep = source.dependencies?.[package]
+//         const devDep = source.devDependencies?.[package]
+//         const peerDep = source.devDependencies?.[package]
+//         const optionalDep = source.devDependencies?.[package]
 
-    for (const package of params.packages) {
-        const dep = source.dependencies?.[package]
-        const devDep = source.devDependencies?.[package]
-        const peerDep = source.devDependencies?.[package]
-        const optionalDep = source.devDependencies?.[package]
+//         updatedDeps = {
+//             ...updatedDeps,
+//             dependencies: {
+//                 ...(updatedDeps.dependencies ?? {}),
+//                 ...(!!dep && { [package]: dep })
+//             },
+//             devDependencies: {
+//                 ...(updatedDeps.devDependencies ?? {}),
+//                 ...(!!devDep && { [package]: devDep })
+//             },
+//             peerDependencies: {
+//                 ...(updatedDeps.peerDependencies ?? {}),
+//                 ...(!!peerDep && { [package]: peerDep })
+//             },
+//             optionalDependencies: {
+//                 ...(updatedDeps.optionalDependencies ?? {}),
+//                 ...(!!optionalDep && { [package]: optionalDep })
+//             }
+//         }
+//     }
 
-        updatedDeps = {
-            ...updatedDeps,
-            dependencies: {
-                ...(updatedDeps.dependencies ?? {}),
-                ...(!!dep && { [package]: dep })
-            },
-            devDependencies: {
-                ...(updatedDeps.devDependencies ?? {}),
-                ...(!!devDep && { [package]: devDep })
-            },
-            peerDependencies: {
-                ...(updatedDeps.peerDependencies ?? {}),
-                ...(!!peerDep && { [package]: peerDep })
-            },
-            optionalDependencies: {
-                ...(updatedDeps.optionalDependencies ?? {}),
-                ...(!!optionalDep && { [package]: optionalDep })
-            }
-        }
-    }
+//     destination.dependencies = {
+//         ...destination?.dependencies,
+//         ...updatedDeps.dependencies
+//     }
+//     destination.devDependencies = {
+//         ...destination?.devDependencies,
+//         ...updatedDeps.devDependencies
+//     }
+//     destination.peerDependencies = {
+//         ...destination?.peerDependencies,
+//         ...updatedDeps.peerDependencies
+//     }
+//     destination.optionalDependencies = {
+//         ...destination?.optionalDependencies,
+//         ...updatedDeps.optionalDependencies
+//     }
 
-    destination.dependencies = {
-        ...destination?.dependencies,
-        ...updatedDeps.dependencies
-    }
-    destination.devDependencies = {
-        ...destination?.devDependencies,
-        ...updatedDeps.devDependencies
-    }
-    destination.peerDependencies = {
-        ...destination?.peerDependencies,
-        ...updatedDeps.peerDependencies
-    }
-    destination.optionalDependencies = {
-        ...destination?.optionalDependencies,
-        ...updatedDeps.optionalDependencies
-    }
+//     fs.writeFileSync(`${params.destination}/package.json`, JSON.stringify(destination, null, 4))
 
-    fs.writeFileSync(`${params.destination}/package.json`, JSON.stringify(destination, null, 4))
+//     // Updating package-lock.json
+//     yellowText("Updating package-lock.json ...")
+//     await execProm(`cd ${params.destination} && npm prune`)
+//     greenText("package-lock.json updated!")
 
-    // Updating package-lock.json
-    yellowText("Updating package-lock.json ...")
-    execSync(`cd ${params.destination} && npm prune`)
-    greenText("package-lock.json updated!")
-
-}
+// }
 
 async function copyFolders(names) {
     // Make the main directories
     let package_names = params.packages.map(el => `${params.source}/node_modules/${el}`).join(' ') + ' ' + names.map(el => `${params.source}/node_modules/${el}`).join(' ')
 
-    const cmd = `mkdir -p ${params.destination}/node_modules`
+    let cmd = ``
     let cmd2 = ``;
-    if(isWindows()) {
+    if (isWindows()) {
+        cmd = `if not exist "${path.resolve(params.destination)}/node_modules mkdir ${path.resolve(params.destination)}/node_modules`
         cmd2 = `pm-copy ${params.destination} ${package_names}`
-    }else{
+    } else {
+        cmd = `mkdir -p ${params.destination}/node_modules`
         cmd2 = `rsync --ignore-missing-args -r ${package_names} ${params.destination}/node_modules`
     }
 
-    exec(`${cmd} && ${cmd2}`);
+    try {
+        await execProm(`${cmd} && ${cmd2}`);
+    } catch (e) {
+        redText(`[Error] Failed to copy the packages\nActual error: ${e.message}`)
+        process.exit(1)
+    }
 
 }